[MODIFY] dao 통합 및 test파일 수정
