+++ src/models/user.dao.js
@@ -1,4 +1,4 @@
-const { AppData } = require("./dataSource");
+const AppData = require("./dataSource");
 
 const getUserById = async (userId) => {
   const [user] = await AppData.query(
@@ -13,4 +13,32 @@ const getUserById = async (userId) => {
   return user;
 };
 
-module.exports = { getUserById };
+const signUp = async (email, kakaoId) => {
+  return AppData.query(
+    `
+    INSERT INTO users (
+        
+        email,
+        kakao_id
+      ) VALUES (?,?)`,
+
+    [email, kakaoId]
+  );
+};
+
+const getUserByEmail = async (email) => {
+  const [users] = await AppData.query(
+    `
+        SELECT
+            
+            u.email,
+            u.kakao_id
+        FROM users u
+        WHERE u.email = ?;
+        `,
+    [email]
+  );
+  return users;
+};
+
+module.exports = { getUserById, signUp, getUserByEmail };
